+++ assets/app/a2services/project-service.js
@@ -109,8 +109,11 @@ angular.module('a2.srv.project', [
                 });
 
                 var serializedParams = $httpParamSerializer(params);
-
-                return '/api/project/'+url+'/recordings/'+ (projection && projection.species ? 'occupancy-models-export/'+projection.species_name+'.csv' : 'recordings-export.csv') + (serializedParams.length ? '?'+serializedParams : '');
+                console.log('projection', projection);
+                return '/api/project/'+url+'/recordings/'+ (
+                    projection && projection.species ? 'occupancy-models-export/'+projection.species_name+'.csv' :
+                    (projection && projection.grouped ? 'grouped-detections-export.csv' : 'recordings-export.csv'))
+                    + (serializedParams.length ? '?'+serializedParams : '');
             },
             getRecTotalQty: function(callback) {
                 return a2APIService.get('/recordings/count').then(function(data) {